Refactor Token
