+++ lib/google_apis/generator/elixir_generator.ex
@@ -1,4 +1,4 @@
-# Copyright 2017 Google Inc.
+# Copyright 2019 Google Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -29,66 +29,9 @@ defmodule GoogleApis.Generator.ElixirGenerator do
     Property,
     Renderer,
     ResourceContext,
+    Token,
     Type
   }
-
-  defmodule Token do
-    defstruct [
-      :filename,
-      :library_name,
-      :namespace,
-      :base_dir,
-      :rest_description,
-      :models_by_name
-    ]
-
-    def build(api_config) do
-      filename = ApiConfig.file(api_config)
-      library_name = ApiConfig.library_name(api_config)
-      namespace = ApiConfig.library_namespace(api_config)
-
-      base_dir =
-        Path.join([
-          "clients",
-          library_name,
-          "lib",
-          "google_api",
-          library_name,
-          String.downcase(ApiConfig.module_version(api_config))
-        ])
-
-      rest_description =
-        api_config
-        |> ApiConfig.google_spec_file()
-        |> File.read!()
-        |> Poison.decode!(as: %GoogleApi.Discovery.V1.Model.RestDescription{})
-
-      %__MODULE__{
-        filename: filename,
-        library_name: library_name,
-        namespace: namespace,
-        base_dir: base_dir,
-        rest_description: rest_description
-      }
-    end
-
-    def build_property(token, model, name, schema) do
-      type =
-        Type.from_schema(schema, %ResourceContext{
-          namespace: token.namespace,
-          model: model,
-          property: name
-        })
-
-      %Property{
-        name: name,
-        description: schema.description,
-        required: schema.required,
-        default: schema.default,
-        type: type
-      }
-    end
-  end
 
   def generate_client(api_config) do
     Token.build(api_config)
